requirements.txt
# Web framework
flask>=3.0.0
flask-socketio>=5.3.0
eventlet>=0.35.0

# Environment management
python-dotenv>=1.0.0

# OpenAI (for document parsing / field extraction)
openai>=1.12.0

# Browser automation
playwright>=1.40.0

# Data processing
pandas>=2.1.0

# Web scraping / parsing
beautifulsoup4>=4.12.0
requests>=2.31.0

# Document parsing
python-docx>=1.1.0
openpyxl>=3.1.0
pdfplumber>=0.11.0

# Logging
colorlog>=6.8.0

# Itinerary analysis (PDF extraction via PyMuPDF)
PyMuPDF>=1.24.0

# Web research (optional - set EXA_API_KEY to enable)
exa-py>=1.0.18
